This change makes it possible to use Imgur with the default browser, instead of the embedded which causes many issues. Other plugins need to follow.
