+++ GreenshotPhotobucketPlugin/PhotobucketUtils.cs
@@ -24,6 +24,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Xml;
 using GreenshotPlugin.Core;
+using GreenshotPlugin.Core.OAuth;
 using GreenshotPlugin.IniFile;
 using GreenshotPlugin.Interfaces;
 using GreenshotPlugin.Interfaces.Plugin;
